requirements.txt
psycopg2-binary==2.9.9
pytest==8.1.1
coverage==7.4.4
